updates, slight simplify
