+++ llava/generate.py
@@ -2,10 +2,12 @@
 
 import argparse
 import codecs
+from pathlib import Path
 
 import mlx.core as mx
+import requests
+from PIL import Image
 from transformers import AutoProcessor
-from utils import get_model_path, load_image, prepare_inputs
 
 from llava import LlavaModel
 
@@ -44,10 +46,42 @@ def parse_arguments():
     return parser.parse_args()
 
 
+def load_image(image_source):
+    """
+    Helper function to load an image from either a URL or file.
+    """
+    if image_source.startswith(("http://", "https://")):
+        try:
+            response = requests.get(image_source, stream=True)
+            response.raise_for_status()
+            return Image.open(response.raw)
+        except Exception as e:
+            raise ValueError(
+                f"Failed to load image from URL: {image_source} with error {e}"
+            )
+    elif Path(image_source).is_file():
+        try:
+            return Image.open(image_source)
+        except IOError as e:
+            raise ValueError(f"Failed to load image {image_source} with error: {e}")
+    else:
+        raise ValueError(
+            f"The image {image_source} must be a valid URL or existing file."
+        )
+
+
+def prepare_inputs(processor, image, prompt):
+    if isinstance(image, str):
+        image = load_image(image)
+    inputs = processor(prompt, image, return_tensors="np")
+    pixel_values = mx.array(inputs["pixel_values"])
+    input_ids = mx.array(inputs["input_ids"])
+    return input_ids, pixel_values
+
+
 def load_model(model_path):
     processor = AutoProcessor.from_pretrained(model_path)
-
-    model = LlavaModel.from_pretrained(get_model_path(model_path))
+    model = LlavaModel.from_pretrained(model_path)
     return processor, model
 
 
@@ -79,12 +113,11 @@ def generate_text(input_ids, pixel_values, model, processor, max_tokens, tempera
 
 def main():
     args = parse_arguments()
-    image = load_image(args.image)
     processor, model = load_model(args.model)
 
     prompt = codecs.decode(args.prompt, "unicode_escape")
 
-    input_ids, pixel_values = prepare_inputs(processor, image, prompt)
+    input_ids, pixel_values = prepare_inputs(processor, args.image, prompt)
 
     print(prompt)
     generated_text = generate_text(
